+++ pyproject.toml
@@ -1,28 +1,46 @@
 [project]
 name = "yapper"
-version = "0.2.0b1"
-description = "Simple python parser converting python docstrings to .astro files for the Astro static site generator."
-authors = [
-    {name = "Gareth Simons", email = "garethsimons@me.com"},
-]
-license-expression = "MIT"
-dependencies = ["docstring-parser<1.0,>=0.13", "PyYAML<7.0,>=6.0", "dominate<3.0.0,>=2.6.0", "python-slugify<7.0.0,>=6.1.1", "Markdown<4.0.0,>=3.3.6"]
+version = "0.2.0b11"
+description = "Parser for converting python docstrings to .astro files for the Astro static site generator."
+readme = 'README.md'
 requires-python = ">=3.9,<4.0"
-
-[project.urls]
-Documentation =  "https://github.com/benchmark-urbanism/yapper"
-Issues = "https://github.com/benchmark-urbanism/yapper/issues"
-
-[project.optional-dependencies]
+license-expression = 'MIT'
+keywords = ['python', 'static-site-generator', 'astro', 'parser', 'documentation', 'docstrings']
+authors = [
+    {name = "Gareth Simons", email = "info@benchmarkurbanism.com"},
+]
+maintainers = [
+    {name = "Gareth Simons", email = "info@benchmarkurbanism.com"},
+]
+classifiers = [
+  "Programming Language :: Python"
+]
+dependencies = [
+    "docstring-parser<1.0,>=0.13",
+    "dominate<3.0.0,>=2.6.0",
+    "python-slugify<7.0.0,>=6.1.1",
+    "Markdown<4.0.0,>=3.3.6",
+    "pyyaml>=6.0",
+]
 [tool.pdm]
+[project.optional-dependencies]
 [tool.pdm.dev-dependencies]
 dev = [
     "pytest>=7.1.1",
 ]
 
+[project.urls]
+homepage =  "https://github.com/benchmark-urbanism/yapper"
+documentation =  "https://github.com/benchmark-urbanism/yapper"
+repository = "https://github.com/benchmark-urbanism/yapper"
+
 [build-system]
-requires = ["pdm-pep517>=0.12.0"]
+requires = ["pdm-pep517>=0.12.0", "tomli>=1.1.0"]
 build-backend = "pdm.pep517.api"
 
 [project.scripts]
-mycli = "yapper.__init__:cli"
+yapper = "yapper.runner:cli"
+
+[tool.pdm.scripts]
+tests = 'pytest ./tests'
+publish = 'pdm publish -u shongololo -v'
