+++ tests/.yap_config_basic.yaml
@@ -1,4 +1,4 @@
-package_root_relative_path: '..'
+package_root_relative_path: './'
 module_map:
   test.mock_file:
     py: ./tests/mock_file.py
